refactor: consolidate ISM compatibility logic and static deployment checks
